refactor: extract shared create_identity helper for RLN binaries
create_identity was near-duplicated across run_rln_proof and register_member,
differing only in return-tuple order and whether the hex-encoded secret hash
was produced. Moves the canonical implementation into rln::client returning
a named-field RlnIdentity struct, and updates both binaries to consume it.

Also drops the multi-line narrative docstring at the top of run_rln_proof
in favor of a single-line description.
